feat(server): migrate /api/plugins/memory/audit to Bun-shape handler (#147 TB11)
