+++ server.ts
@@ -53,6 +53,7 @@ import * as pluginSettingsIdRoute from './packages/host/src/api/plugin-settings/
 import * as pluginSettingsSchemasRoute from './packages/host/src/api/plugin-settings/schemas'
 import * as pluginsInstallRoute from './packages/host/src/api/plugins/install'
 import * as pluginsRemoveRoute from './packages/host/src/api/plugins/remove'
+import * as pluginsMemoryAuditRoute from './packages/host/src/api/plugins/memory/audit'
 
 const log = createLogger('server')
 
@@ -410,6 +411,13 @@ app.prepare().then(async () => {
 
     if (url.pathname === '/api/plugin-settings/schemas' && req.method === 'GET') {
       dispatchWebHandler(req, res, pluginSettingsSchemasRoute.get)
+      return
+    }
+
+    // /api/plugins/memory/{audit,gateway,workspace} — former standalone
+    // routes now served directly (they shadow the plugin catch-all).
+    if (url.pathname === '/api/plugins/memory/audit' && req.method === 'GET') {
+      dispatchWebHandler(req, res, pluginsMemoryAuditRoute.get)
       return
     }
